Implemented auth-middleware
